Move chat code to separate files
